.antigravity/scratch/gemini_assistant/requirements.txt
google-generativeai
SpeechRecognition
pyttsx3
pyaudio
pyautogui
python-dotenv
